标签分类搞定

+++ blog-boot/src/main/java/com/dkt/blogboot/service/ArticleService.java
@@ -2,6 +2,7 @@ package com.dkt.blogboot.service;
 
 import com.dkt.blogboot.entity.Article;
 import com.dkt.blogboot.entity.ArticleCategory;
+import com.dkt.blogboot.entity.ResponseBean;
 import com.dkt.blogboot.mapper.ArticleCategoryMapper;
 import com.dkt.blogboot.mapper.ArticleMapper;
 import com.dkt.blogboot.mapper.ArticleTagMapper;
@@ -18,7 +19,7 @@ import java.util.List;
  * @Date 2020-08-05 15:22
  */
 @Service
-@Transactional
+@Transactional(rollbackFor = Exception.class)
 public class ArticleService {
 
     @Autowired
@@ -28,22 +29,36 @@ public class ArticleService {
     @Autowired
     ArticleTagMapper articleTagMapper;
 
-    public int addArticle(Article article) {
+    public ResponseBean addArticle(Article article) {
         if (article.getId() == -1) {
             article.setDate(new Date(new java.util.Date().getTime()));
+            if (article.getTagIdArr().length <= 0 || article.getCategoryId() <= 0 || "".equals(article.getContent()) || article.getContent() == null || "".equals(article.getTitle()) || article.getTitle() == null) {
+                return new ResponseBean("error", "标题、内容、分类、标签存在空值，添加失败");
+            }
             int insert = articleMapper.insert(article);
             ArticleCategory articleCategory = new ArticleCategory(article.getId(), article.getCategoryId());
             int insert1 = articleCategoryMapper.insert(articleCategory);
             int inserts = articleTagMapper.inserts(article.getId(), article.getTagIdArr());
-            return insert;
+            if (insert >= 1 && insert1 >= 1 && inserts >= 1) {
+                return new ResponseBean("success", "添加文章成功");
+            } else {
+                return new ResponseBean("error", "添加文章失败");
+            }
         } else {
+            if (article.getTagIdArr().length <= 0 || article.getCategoryId() <= 0 || "".equals(article.getContent()) || article.getContent() == null || "".equals(article.getTitle()) || article.getTitle() == null) {
+                return new ResponseBean("error", "标题、内容、分类、标签存在空值，添加失败");
+            }
             int updateArticle = articleMapper.updateByPrimaryKey(article);
             int deleteByAid = articleCategoryMapper.deleteByAid(article.getId());
             ArticleCategory articleCategory = new ArticleCategory(article.getId(), article.getCategoryId());
             int insert = articleCategoryMapper.insert(articleCategory);
             int deleteByAid1 = articleTagMapper.deleteByAid(article.getId());
             int inserts = articleTagMapper.inserts(article.getId(), article.getTagIdArr());
-            return updateArticle;
+            if (updateArticle >= 1 && insert >= 1 && inserts >= 1) {
+                return new ResponseBean("success", "修改文章成功");
+            } else {
+                return new ResponseBean("error", "修改文章失败");
+            }
         }
     }
 
@@ -76,11 +91,15 @@ public class ArticleService {
         return formatDate(article);
     }
 
-    public int deleteArticleById(int id) {
+    public ResponseBean deleteArticleById(int id) {
         int deleteByAid = articleCategoryMapper.deleteByAid(id);
         int deleteByAid1 = articleTagMapper.deleteByAid(id);
         int deleteByPrimaryKey = articleMapper.deleteByPrimaryKey(id);
-        return deleteByPrimaryKey;
+        if (deleteByAid >= 1 && deleteByAid1 >= 1 && deleteByPrimaryKey >= 1) {
+            return new ResponseBean("success", "删除文章成功");
+        } else {
+            return new ResponseBean("error", "删除文章失败");
+        }
     }
 
     public List<Article> getAllArticleSubstringContent() {
